added more oop
