+++ src/main/java/BenBenException.java
@@ -1,0 +1,10 @@
+class BenBenException extends RuntimeException {
+    BenBenException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return "WOOF! " + getMessage();
+    }
+}
