+++ src/main/java/tasks.txt
@@ -1,3 +1,2 @@
-E | 1 | meeting | 2023-08-31 00:00 | 2023-09-01 00:00
-T | 0 | ip
-E | 0 | trip | 2023-12-01 23:59 | 2023-12-31 12:12
+D | 0 | ip | 2023-08-31
+E | 0 | meeting | 2023-09-01 10:00 | 2023-09-01 12:00
